+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added a `showCounts` option to show the number of items on each side of a LeftRightChooser.
 * Added an `useOnResize` hook, which runs a function when a component is resized.
+* Added `StoreFilterField` to the mobile toolkit. Its API is identical to the desktop version.
 
 ### 💥 Breaking Changes